+++ 03_canny/kernels.cpp
@@ -3,7 +3,7 @@
 cv::Mat1f gaussian(float sigma) {
     int a = round(2.5 * sigma - 0.5);
     // printf("preparing kernel of length %d\n", 2*a+1);
-    cv::Mat1f kern(1, 2*a+1, CV_16FC1);
+    cv::Mat1f kern(1, 2*a+1, CV_32FC1);
     float sum = 0;
     for (auto i = kern.begin(); i != kern.end(); i++) {
         *i = exp((-1.0 * (i.pos().x-a) * (i.pos().x-a)) / (2.0 * sigma * sigma));
@@ -18,7 +18,7 @@ cv::Mat1f gaussian(float sigma) {
 cv::Mat1f deriv(float sigma) {
     int a = round(2.5 * sigma - 0.5);
     // printf("preparing deriv kernel of length %d\n", 2*a+1);
-    cv::Mat1f kern(1, 2*a+1, CV_16FC1);
+    cv::Mat1f kern(1, 2*a+1, CV_32FC1);
     float sum = 0;
     for (auto i = kern.begin(); i != kern.end(); i++) {
         *i = -1.0 * (i.pos().x-a) * exp((-1.0 * (i.pos().x-a) * (i.pos().x-a)) / (2.0 * sigma * sigma));
